refactor(geometry): carve rtc_offset + diagnostics out of router/{processing,mod}
router/processing.rs (1430): the ~380-LOC RTC-offset detection block
(detect_rtc_offset_* + sample_*/first_vertex helpers) moves to a sibling
rtc_offset.rs as a second impl GeometryRouter block (processing.rs now 1060,
the remaining cohesive meshing orchestrator).

router/mod.rs (785): the ~300-LOC telemetry layer (ClassificationStats /
*Diagnostic types + record_*/take_*/bump_* accumulators) moves to a sibling
diagnostics.rs (mod.rs now 487). The child module reads GeometryRouter's
private RefCell fields without field-visibility changes.

Pure code motion; one pub(super) bump (representation_item_uses_raw_large_
coordinates). Re-exports keep lib.rs:142-144 + voids.rs paths identical;
lib.rs untouched. build/clippy -D warnings clean; rtc/router tests green.
